Updates to PLOS version of manuscript
I believe I have moved over all relevant text to the PLOS version of the
manuscript. There are a few in-text references to subsections (and maybe
to other things like equations that may not be working). I may also need
to reformat the equations into PLOS style. I am going to wait to fix
this stuff until I start working on the text again since some stuff
might move out of the Methods and into an appendix.

I've compiled the references and they are in the document, though before
submission we have to actually paste the contents of the .bbl file into
the .tex file.